# docs/swagger.yaml
basePath: /
definitions:
  entity.Checkout:
    properties:
      checkout:
        items:
          $ref: '#/definitions/entity.CheckoutRequest'
        type: array
    type: object
  entity.CheckoutRequest:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
...
info:
  contact: {}
  title: Kasir API
  version: "1.0"
paths:
...
  /api/reports:
    get:
      consumes:
      - application/json
      description: Get sales report with or without Date Range (Default Today)
      parameters:
      - description: your-secret-api-key-here
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Start Date
        in: query
        name: start_date
        type: string
      - description: End Date
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get sales report with or without Date Range (Default Today)
      tags:
      - reports
  /api/reports/hari-ini:
    get:
      consumes:
      - application/json
      description: Get sales report daily
      parameters:
      - description: your-secret-api-key-here
        in: header
        name: X-API-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get sales report daily
      tags:
      - reports
...
  /api/transactions/checkout:
    post:
      consumes:
      - application/json
      description: Checkout products
      parameters:
      - description: your-secret-api-key-here
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Checkout Data
        in: body
        name: checkout
        required: true
        schema:
          $ref: '#/definitions/entity.Checkout'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Checkout products
      tags:
      - transactions
...
swagger: "2.0"